Правки в контроллере. Изменен формат поля passport в сущности Student.

+++ src/main/java/it/university/testapp/StudentsController.java
@@ -3,6 +3,7 @@ package it.university.testapp;
 import it.university.testapp.entity.Student;
 import it.university.testapp.service.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -20,9 +21,12 @@ public class StudentsController {
             produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
     public ResponseEntity<Student> addStudent(@RequestBody Student student) {
+        if (!studentService.passportIsValid(student)) {
+            return ResponseEntity.badRequest().build();
+        }
         Student resultStudent = studentService.addStudent(student);
         if (resultStudent == null) {
-            return ResponseEntity.badRequest().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         } else {
             return ResponseEntity.ok().body(student);
         }
@@ -33,9 +37,15 @@ public class StudentsController {
             produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
     public ResponseEntity<Student> updateStudent(@PathVariable("studentId") int studentId, @RequestBody Student student) {
+        if (!studentService.passportIsValid(student)) {
+            return ResponseEntity.badRequest().build();
+        }
+        if (!studentService.studentExists(studentId)) {
+            return ResponseEntity.notFound().build();
+        }
         Student resultStudent = studentService.updateStudentById(student, studentId);
         if (resultStudent == null) {
-            return ResponseEntity.badRequest().build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
         } else {
             return ResponseEntity.ok().body(student);
         }
@@ -47,7 +57,7 @@ public class StudentsController {
     public ResponseEntity<Student> getStudent(@PathVariable("studentId") int studentId) {
         Student resultStudent = studentService.getStudent(studentId);
         if (resultStudent == null) {
-            return ResponseEntity.noContent().build();
+            return ResponseEntity.notFound().build();
         } else {
             return ResponseEntity.ok().body(resultStudent);
         }
@@ -64,7 +74,11 @@ public class StudentsController {
             method = RequestMethod.DELETE)
     @ResponseBody
     public ResponseEntity<Student> deleteStudent(@PathVariable("studentId") int studentId) {
-        studentService.deleteStudent(studentId);
-        return ResponseEntity.ok().build();
+        if (!studentService.studentExists(studentId)) {
+            return ResponseEntity.notFound().build();
+        } else {
+            studentService.deleteStudent(studentId);
+            return ResponseEntity.ok().build();
+        }
     }
 }
